SilentSignal/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
boto3==1.39.0
httpx==0.28.1
